# whatsudow_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatsudow",
        "description": "Aplicación que permite a los organizadores de eventos unificar los datos de sus proveedores de servicios. Con esta app, los organizadores de eventos pueden almacenar y acceder a información importante de los proveedores, comocalendarios de disponibilidad, PDF actualizados, cláusulas y datos de contacto.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/budgets": {
            "get": {
                "tags": [
                    "Budgets"
                ],
                "summary": "Muestra los presupuestos especificos",
                "description": "Display a listing of the resource.",
                "operationId": "getBudgets",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    }
                }
            },
            "post": {
                "tags": [
                    "Budgets"
                ],
                "summary": "Create a new budget",
                "description": "Store a newly created resource in storage.",
                "operationId": "storeBudget",
                "requestBody": {
                    "description": "Almacena un nuevo presupuesto de un usuario concreto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "event_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Presupuesto creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/budgets/{id}": {
            "get": {
                "tags": [
                    "Budgets"
                ],
                "summary": "Obtener detalles de un presupuesto específico",
                "description": "Display the specified resource.",
                "operationId": "showBudget",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de presupuesto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "budget_title": {
                                            "type": "string"
                                        },
                                        "sender_name": {
                                            "type": "string"
                                        },
                                        "received_date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "sent_date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "message": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "event_date": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "location": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Budgets"
                ],
                "summary": "Eliminar un presupuesto específico",
                "description": "Remove the specified resource from storage.",
                "operationId": "destroyBudget",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de presupuesto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Presupuesto eliminado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mensaje eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}